Initial edits to use double-precision data.
 The 3.x branch is in development and is currently
 NOT functional -- 2025-03-29 km.

 On branch 3.x
	modified:   modules/xyplot.4th
	modified:   src/AlgebraCompiler.cpp
	modified:   src/CDatabase.cpp
	modified:   src/CDatabase.h
	modified:   src/CDataset.cpp
	modified:   src/CDataset.h
	modified:   src/CDeviceContext.cpp
	modified:   src/CGrid.cpp
	modified:   src/CPlotDisplay.cpp
	modified:   src/CPlotDisplay.h
	modified:   src/CPlotList.cpp
	modified:   src/CPlotList.h
	modified:   src/CPlotView.cpp
	modified:   src/CPlotView.h
	modified:   src/CPlotWindow.cpp
	modified:   src/CPolarGrid.cpp
	modified:   src/CTransform.cpp
	modified:   src/CTransform.h
	modified:   src/CXyFile.cpp
	modified:   src/CXyFile.h
	modified:   src/CXyPlot.cpp
	modified:   src/CpsDC.cpp
	modified:   src/CpsDC.h
	modified:   src/Makefile
	modified:   src/xyplot.cpp
	modified:   src/xyplot.h

+++ src/AlgebraCompiler.cpp
@@ -1,6 +1,6 @@
 // AlgebraCompiler.cpp
 //
-// Copyright (c) 1998--2020 Krishna Myneni
+// Copyright (c) 1998--2023 Krishna Myneni
 //
 // This software is provided under the terms of the
 // GNU Affero General Public License (AGPL) v 3.0 or later.
@@ -41,7 +41,7 @@ int HoldOperator( byte, stack<byte>*, deque<byte>*, int );
 //      addr  nsize  npts
 //
 // where addr is the starting address of the data, nsize is the
-// number of elements (floats) in a datum, and npts is the number
+// number of elements (doubles) in a datum, and npts is the number
 // of points.
 //
 // The stack diagram for an expression evaluation is:
@@ -85,11 +85,11 @@ int CompileAE (vector<byte>* pOpCodes, char* exp)
     // Set up the loop parameters
     op.push_back(OP_SWAP);      // bring nsize to top
     op.push_back(OP_IVAL);
-    ival = sizeof(float);
+    ival = sizeof(double);
     bp = (byte*)&ival;
     for (i = 0; i < sizeof(int); i++)
         op.push_back(*(bp + i));
-    op.push_back(OP_MUL);       // multiply nsize by sizeof(float)
+    op.push_back(OP_MUL);       // multiply nsize by sizeof(double)
     op.push_back(OP_SWAP);      // bring npts to top
     op.push_back(OP_IVAL);
     for (i = 0; i < sizeof(int); i++) op.push_back(0);
